requirements.txt
yt-dlp==2023.12.30
gallery-dl==1.26.5
python-dotenv==1.0.1